Updating starterkit to make scenarios easier to customise
